# src/main/java/it/gov/pagopa/pu/processecexutions/controller/IngestionFlowFileEntityExtendedController.java
package it.gov.pagopa.pu.processecexutions.controller;

import it.gov.pagopa.pu.processecexutions.controller.generated.IngestionFlowFileEntityExtendedControllerApi;
import it.gov.pagopa.pu.processecexutions.enums.IngestionFlowFileStatus;
import it.gov.pagopa.pu.processecexutions.repository.IngestionFlowFileRepository;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

/**
 * Controller to host spring-data-rest directly not supported methods
 */
@RestController
public class IngestionFlowFileEntityExtendedController implements IngestionFlowFileEntityExtendedControllerApi {

  private final IngestionFlowFileRepository repository;

  public IngestionFlowFileEntityExtendedController(IngestionFlowFileRepository repository) {
    this.repository = repository;
  }

  @Override
  public ResponseEntity<Integer> updateIngestionFlowFileStatus(Long ingestionFlowFileId, IngestionFlowFileStatus oldStatus, IngestionFlowFileStatus newStatus,
                                              String errorDescription,String discardFile) {
    int result = repository.updateStatus(ingestionFlowFileId, oldStatus, newStatus, errorDescription, discardFile);
    if(result>0){
      return ResponseEntity.ok(result);
    } else {
      return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
    }
  }
}